Rename imagePath package to path and move DirectoryPath inside
