+++ docs/DeveloperGuide.adoc
@@ -14,7 +14,7 @@ ifdef::env-github[]
 endif::[]
 :repoURL: https://github.com/AY1920S1-CS2103-T16-3/main
 
-By: `Team SE-EDU`      Since: `Jun 2016`      Licence: `MIT`
+By: `CS2103-T16-3`      Since: `September 2019`      Licence: `MIT`
 
 == Setting up
