+++ src/main/java/seedu/weme/logic/commands/createcommand/CreateCommand.java
@@ -12,9 +12,9 @@ import seedu.weme.logic.commands.exceptions.CommandException;
 import seedu.weme.model.Model;
 import seedu.weme.model.ModelContext;
 import seedu.weme.model.ReadOnlyUserPrefs;
-import seedu.weme.model.imagePath.ImagePath;
 import seedu.weme.model.meme.Description;
 import seedu.weme.model.meme.Meme;
+import seedu.weme.model.path.ImagePath;
 import seedu.weme.model.tag.Tag;
 import seedu.weme.model.template.exceptions.MemeCreationException;
 import seedu.weme.model.util.ImageUtil;
