+++ src/main/java/seedu/weme/logic/commands/importcommand/LoadCommand.java
@@ -11,8 +11,8 @@ import seedu.weme.commons.util.FileUtil;
 import seedu.weme.logic.commands.Command;
 import seedu.weme.logic.commands.CommandResult;
 import seedu.weme.logic.commands.exceptions.CommandException;
-import seedu.weme.model.DirectoryPath;
 import seedu.weme.model.Model;
+import seedu.weme.model.path.DirectoryPath;
 
 /**
  * Loads memes from a directory to the import context.
