+++ src/test/java/seedu/weme/model/meme/ImagePathTest.java
@@ -6,7 +6,7 @@ import static seedu.weme.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.weme.model.imagePath.ImagePath;
+import seedu.weme.model.path.ImagePath;
 
 public class ImagePathTest {
