+++ src/test/java/seedu/weme/storage/JsonAdaptedMemeTest.java
@@ -13,8 +13,8 @@ import java.util.stream.Collectors;
 import org.junit.jupiter.api.Test;
 
 import seedu.weme.commons.exceptions.IllegalValueException;
-import seedu.weme.model.imagePath.ImagePath;
 import seedu.weme.model.meme.Description;
+import seedu.weme.model.path.ImagePath;
 
 public class JsonAdaptedMemeTest {
     private static final String INVALID_TAG = "#friend";
